Structure of the app with Header, Body, Footer and Routes
Hi Team, I was finally able to push this up to GitHub. I created the structure that we can work with. Each person can focus on a component. Look over the code and try to get familiar with it. It's not 100% done since I still need to work with the API and display the data on the page.